+++ demos/chinook/src/main/java/is/codion/demos/chinook/ui/PlaylistTrackEditPanel.java
@@ -40,7 +40,7 @@ final class PlaylistTrackEditPanel extends EntityEditPanel {
 	@Override
 	protected void initializeUI() {
 		initialFocusAttribute().set(PlaylistTrack.TRACK_FK);
-		createForeignKeySearchField(PlaylistTrack.TRACK_FK)
+		createSearchField(PlaylistTrack.TRACK_FK)
 						.selectorFactory(new TrackSelectorFactory())
 						.transferFocusOnEnter(false)
 						.columns(25);