+++ demos/chinook/src/main/java/is/codion/demos/chinook/ui/TrackEditPanel.java
@@ -48,14 +48,14 @@ public final class TrackEditPanel extends EntityEditPanel {
 	protected void initializeUI() {
 		initialFocusAttribute().set(Track.ALBUM_FK);
 
-		createForeignKeySearchField(Track.ALBUM_FK);
+		createSearchField(Track.ALBUM_FK);
 		createTextField(Track.NAME)
 						.columns(12);
-		createForeignKeyComboBoxPanel(Track.MEDIATYPE_FK, this::createMediaTypeEditPanel)
+		createComboBoxPanel(Track.MEDIATYPE_FK, this::createMediaTypeEditPanel)
 						.preferredWidth(160)
 						.includeAddButton(true)
 						.includeEditButton(true);
-		createForeignKeyComboBoxPanel(Track.GENRE_FK, this::createGenreEditPanel)
+		createComboBoxPanel(Track.GENRE_FK, this::createGenreEditPanel)
 						.preferredWidth(160)
 						.includeAddButton(true)
 						.includeEditButton(true);