+++ demos/employees/src/main/java/is/codion/demos/employees/ui/EmployeeEditPanel.java
@@ -47,7 +47,7 @@ public class EmployeeEditPanel extends EntityEditPanel {
 		createTextField(Employee.NAME)
 						.columns(8);
 		createItemComboBox(Employee.JOB);
-		createForeignKeyComboBox(Employee.MANAGER_FK);
+		createComboBox(Employee.MANAGER_FK);
 		createTextField(Employee.SALARY)
 						.columns(5);
 		createTextField(Employee.COMMISSION)
@@ -70,7 +70,7 @@ public class EmployeeEditPanel extends EntityEditPanel {
 	}
 
 	private JPanel createDepartmentPanel() {
-		EntityComboBox departmentBox = createForeignKeyComboBox(Employee.DEPARTMENT_FK).build();
+		EntityComboBox departmentBox = createComboBox(Employee.DEPARTMENT_FK).build();
 		NumberField<Integer> departmentNumberField = departmentBox.integerSelectorField(Department.DEPARTMENT_NO)
 						.transferFocusOnEnter(true)
 						.onBuild(field -> addValidator(Employee.DEPARTMENT, field))