+++ changelog.md
@@ -110,6 +110,7 @@ Codion Change Log
 - EntityPanel.activate() bug fixed, requests display first, to ensure correct initial focus.
 - EntityPanel.Display change reverted, activate() and activated() reintroduced, needs polishing.
 - EntityComponents.foreignKeyComboBox(), foreignKeyComboBoxPanel(), foreignKeySearchField(), foreignKeySearchFieldPanel(), foreignKeyTextField() and foreignKeyLabel() renamed comboBox(), comboBoxPanel(), searchField(), searchFieldPanel(), textField() and entityLabel() respectively.
+- EntityEditComponentPanel.createForeignKeyComboBox(), createForeignKeyComboBoxPanel(), createForeignKeySearchField(), createForeignKeySearchFieldPanel(), createForeignKeyTextField() and createForeignKeyLabel() renamed createComboBox(), createComboBoxPanel(), createSearchField(), createSearchFieldPanel(), createTextField() and createEntityLabel() respectively.
 
 ## 0.18.22
 ### is.codion.common.core